+++ maro/simulator/core.py
@@ -8,8 +8,7 @@ from typing import List
 
 from maro.backends.frame import FrameBase, SnapshotList
 from maro.event_buffer import EventBuffer, EventState
-from maro.utils.exception.simulator_exception import \
-  BusinessEngineNotFoundError
+from maro.utils.exception.simulator_exception import BusinessEngineNotFoundError
 
 from .abs_core import AbsEnv, DecisionMode
 from .scenarios.abs_business_engine import AbsBusinessEngine
