+++ maro/simulator/scenarios/cim/business_engine.py
@@ -11,17 +11,17 @@ from maro.backends.frame import FrameBase, SnapshotList
 from maro.data_lib.cim import CimDataContainerWrapper, Order, Stop
 from maro.event_buffer import AtomEvent, CascadeEvent, EventBuffer, MaroEvents
 from maro.simulator.scenarios import AbsBusinessEngine
-from maro.simulator.scenarios.helpers import (
-    DocableDict, MatrixAttributeAccessor)
+from maro.simulator.scenarios.helpers import DocableDict
+from maro.simulator.scenarios.matrix_accessor import MatrixAttributeAccessor
 
-from .common import (
-    Action, ActionScope, DecisionEvent, EmptyReturnPayload,
-    LadenReturnPayload, VesselDischargePayload,
-    VesselStatePayload)
+from .common import Action, ActionScope, DecisionEvent
+from .event_payload import (EmptyReturnPayload, LadenReturnPayload,
+                            VesselDischargePayload, VesselStatePayload)
 from .events import Events
 from .frame_builder import gen_cim_frame
 
-metrics_desc = """
+
+metrics_desc="""
 CIM metrics used provide statistics information until now (may be in the middle of current tick).
 It contains following keys:
