+++ maro/simulator/scenarios/cim/matrix.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT license.
 
 from maro.backends.frame import NodeAttribute, NodeBase, node
-from maro.simulator.scenarios.helpers import MatrixAttributeAccessor
+from maro.simulator.scenarios.matrix_accessor import MatrixAttributeAccessor
 
 
 def gen_matrix(port_num: int, vessel_num: int):
